Enhance formatting and content of bio.html
Linked the page to the master CSS stylesheet.
Added the top toolbar.
Added content regarding interest in physics and the decision to switch to
computer science.
Added "biography" heading to the top.
Created "narration" class to better format the paragraphs, added this class
(which just includes margin adjusting) to master.css, and labelled all
p elements in bio.html with it.