Add picture to index.html
Also added formatting to make the page look nice.  Created a div element to
push the content of the page to the right of the picture instead of below it.
Created a profile-picture class to format the size, as well as to allow it
to be pushed to the left of the page.