Remove styling of some unordered lists
Created a CSS class called "nostyle" and applied it to the list in index.html
and the topmost list in resume.html.  The prior method was to determine how
far down the list was from the body, but this was too general and proved
fragile (it broke when the list was placed in a div element).